move test_reader to _qt/
